add first tests and fixtures
