+++ cg/models/orders/samples.py
@@ -195,6 +195,10 @@ class FastqSample(OrderInSample):
 class PacBioSample(OrderInSample):
     _suitable_project = OrderType.PACBIO_LONG_READ
 
+    capture_kit: str | None
+    sex: SexEnum = SexEnum.unknown
+    tumour: bool
+
 
 class RmlSample(OrderInSample):
     _suitable_project = OrderType.RML
