+++ cg/services/orders/store_order_services/store_pacbio_order_service.py
@@ -50,7 +50,6 @@ class StorePacBioOrderService(StoreOrderService):
                     "name": sample.name,
                     "priority": sample.priority,
                     "sex": sample.sex,
-                    "subject_id": sample.subject_id,
                     "tumour": sample.tumour,
                     "volume": sample.volume,
                 }
@@ -84,7 +83,6 @@ class StorePacBioOrderService(StoreOrderService):
                     priority=sample["priority"],
                     tumour=sample["tumour"],
                     capture_kit=sample["capture_kit"],
-                    subject_id=sample["subject_id"],
                 )
                 new_sample.customer = customer
                 application_tag: str = sample["application"]
