+++ tests/apps/orderform/test_json_orderform_parser.py
@@ -8,11 +8,11 @@ from cg.models.orders.orderform_schema import Orderform
 @pytest.mark.parametrize(
     "valid_json_order_type", [OrderType.MIP_DNA, OrderType.BALSAMIC, OrderType.FLUFFY]
 )
-def test_generate_mip_json_orderform(valid_json_order_type: str, json_order_list):
+def test_generate_mip_json_orderform(valid_json_order_type: str, json_order_dict):
     """Tests the orderform generation for customer-submitted json files"""
 
     # GIVEN a dictionary from a JSON file of a certain order type
-    json_order: dict = json_order_list[valid_json_order_type]
+    json_order: dict = json_order_dict[valid_json_order_type]
 
     # WHEN an orderform is parsed and an Orderform object generated
     order_form_parser = JsonOrderformParser()
