+++ tests/fixture_plugins/orders_fixtures/order_store_service_fixtures.py
@@ -1,8 +1,8 @@
 import pytest
 
 from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
-from cg.services.orders.store_order_services.store_fastq_order_service import StoreFastqOrderService
 from cg.services.orders.store_order_services.store_case_order import StoreCaseOrderService
+from cg.services.orders.store_order_services.store_fastq_order_service import StoreFastqOrderService
 from cg.services.orders.store_order_services.store_metagenome_order import (
     StoreMetagenomeOrderService,
 )
@@ -10,6 +10,9 @@ from cg.services.orders.store_order_services.store_microbial_fastq_order_service
     StoreMicrobialFastqOrderService,
 )
 from cg.services.orders.store_order_services.store_microbial_order import StoreMicrobialOrderService
+from cg.services.orders.store_order_services.store_pacbio_order_service import (
+    StorePacBioOrderService,
+)
 from cg.services.orders.store_order_services.store_pool_order import StorePoolOrderService
 from cg.store.store import Store
 from tests.mocks.limsmock import MockLimsAPI
@@ -28,6 +31,11 @@ def store_pool_order_service(base_store: Store, lims_api: MockLimsAPI) -> StoreP
 @pytest.fixture
 def store_fastq_order_service(base_store: Store, lims_api: MockLimsAPI) -> StoreFastqOrderService:
     return StoreFastqOrderService(status_db=base_store, lims_service=OrderLimsService(lims_api))
+
+
+@pytest.fixture
+def store_pacbio_order_service(base_store: Store, lims_api: MockLimsAPI) -> StorePacBioOrderService:
+    return StorePacBioOrderService(status_db=base_store, lims_service=OrderLimsService(lims_api))
 
 
 @pytest.fixture
