+++ tests/fixtures/cgweb_orders/pacbio.json
@@ -1,0 +1,39 @@
+{
+    "name": "PacbioOrder",
+    "customer": "cust000",
+    "comment": "",
+    "samples": [
+        {
+            "application": "WGSPCFC060",
+            "comment": "",
+            "container": "Tube",
+            "container_name": "prov1",
+            "data_analysis": "raw-data",
+            "data_delivery": "bam",
+            "elution_buffer": "Nuclease-free water",
+            "name": "prov1",
+            "priority": "priority",
+            "sex": "female",
+            "source": "blood",
+            "tumour": false,
+            "volume": "25",
+            "well_position": ""
+        },
+        {
+            "application": "WGSPCFC060",
+            "comment": "",
+            "container": "Tube",
+            "container_name": "prov2",
+            "data_analysis": "raw-data",
+            "data_delivery": "bam",
+            "elution_buffer": "Nuclease-free water",
+            "name": "prov2",
+            "priority": "priority",
+            "sex": "male",
+            "source": "cell line",
+            "tumour": true,
+            "volume": "35",
+            "well_position": ""
+        }
+    ]
+}
